Automated emailing feature implemented
Implemented:
- Emailer Service - responsible for emailing users, using our email templates, stored applicant, company and job application data, and NodeMailer.

To view the sent email, simply click on the provided by NodeMailer link in the backend terminal and it should directly lead you to the sent email.

Feature is implemented using mocked email accounts via Ethereal Email. It's usage is purely for the sake of not creating dummy emails.

Using the given credentials (in the backend terminal again), you can view the history of emails of the currently used Ethereal account. The account lives within the application's life-cycle, meaning each time you restart the application, a new account is made (last one is probably deleted by the Ethereal system).